E-Commerce Website
An e-commerce website is an online platform where goods and services are bought and sold. These websites enable businesses and consumers to conduct transactions electronically via the internet. This is an Puma shoe producing  the Festive shoe robust , Athlet wear and canvases.
This shoe represents the amount of shoe in winter ,summer and normal wear times.